RoomUser updates

+++ Sadie.Game.Rooms/Users/IRoomUserFactory.cs
@@ -9,7 +9,7 @@ public interface IRoomUserFactory
     RoomUser Create(
         IRoom room,
         INetworkObject networkObject, 
-        long id, HPoint point, 
+        int id, HPoint point, 
         HDirection directionHead,
         HDirection direction, 
         IAvatarData avatarData);
